storybook_generator/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
pillow==10.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
aiosqlite==0.19.0
